[][kernel-6.6][mt7988][bsp][Refactor the GPIO reset and increase the MDC frequency to 12.5MHz for the Airoha AN8831X 10G PHY]
[Description]
Refactor the GPIO reset and increase the MDC frequency to 12.5MHz for
the Airoha AN8831X 10G PHY.

In this new RFB, we connect both AN8831X devices using GPIO71.
Therefore, we need to change the GPIO reset from the device level to the
bus level. Additionally, we have increased the MDC frequency to 12.5 MHz
to reduce the time required for firmware downloading.

[Release-log]
N/A


Change-Id: I21c0d64a10cf404001506712e7d7621352b47c8b
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/11089571