[][openwrt][mt7988][crypto][Change the algorithm for hash key precompute from lookaside to sw]
[Description]
Change the algorithm for hash key precompute from lookaside to sw.
To create transform record in EIP197 inline mode, a precomputation for
hash key is necessary.
In the past, we follow DDK examples to perform precompute with EIP197
lookaside mode.
However, we found that when rekey is invoked multiple times in 1 sec,
the precompute with DDK lookaside may lead to cache inconsistency
problem. We now change to use kernel algorithms.

[Release-log]
N/A

Change-Id: If167dc7f0cdfdf7eff1e939fe7bb1b571f31d37d
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/11316627