[][kernel-6.12][common][eth][Add NETSYS SER fast mode]
[Description]
Add NETSYS SER fast mode.

In the current design, NETSYS SER would reset the FE and GMAC in any
hang condition. However, in cases where the GMAC is not hung, we can
use MAC_LINK_DOWN instead of a PHY power-down to save a few seconds of
waiting for the PHY link to come up, thereby avoiding a WiFi traffic
stop.

Without this patch, the traffic between ETH and WiFi would stop for a
few seconds, even if the GMAC is not hung.

[Release-log]
N/A


Change-Id: I77eb6c11e0034a5d36ebd49837e3af2122ed84d6
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/11387873