[][[Coverity][REBB]MISRA C-2012 Standard Libraries: 5604920]
[Description]
Fix mii_mgr.c coverity issue:
The string "eth0" has a size of 5 bytes (including the null terminator),
but IFNAMSIZ - 1 is typically 15 (since IFNAMSIZ is usually 16).

The issue is that strncpy is being asked to copy up to 15 bytes from a
source that only has 5 bytes. While this doesn't cause a runtime error
(strncpy will pad with zeros), it's flagged by Coverity as a potential
issue.

[Release-log]
N/A

Change-Id: Ibf7fc41addd3511a0b2ea5fabf5308bddd2ff266
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/11472772