[][kernel][mt7988][crypto][Refactor skb headroom copy when gso segment]
[Description]
Refactor skb headroom copy when gso segment.
For the skb needs to perform IPSec inline mode, we'll set a mark in
skb headroom.
When perform gso segment, new skbs will create to replace the original
skb and the information in skb headroom will lost.
In the past, this patch will copy the headroom from old skb to the new
one. However, the new skb is a list of segments, we should copy headroom
from the old skb to all segments. This patch fix this issue.

[Release-log]
N/A

Change-Id: I3e43f2173055dfa1f6b20fdd6f32fb98eccc0b08
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/11497340